calculated andprovided feedback on users monthly savings and potential future savings

+++ python_introduction/finance_calculator.py
@@ -1,5 +1,5 @@
-monthly_inc = int(input("Enter your monthly income: "))
-monthly_exp = int(input("Enter your monthly expenses: "))
+monthly_inc = float(input("Enter your monthly income: "))
+monthly_exp = float(input("Enter your monthly expenses: "))
 monthly_sav = monthly_inc - monthly_exp
 print("Your monthly savings are","$",monthly_sav)
 projected_sav = monthly_sav * 12 + (monthly_sav * 12 * 0.05)
